% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{pipe_collapse}
\alias{pipe_collapse}
\title{This function allows you to collapse a list with the pipe separator (useful for select(matches()) subsetting)}
\usage{
pipe_collapse(input_list)
}
\arguments{
\item{input_list}{the list to be collapsed
pipe_collapse()}
}
\description{
This function allows you to collapse a list with the pipe separator (useful for select(matches()) subsetting)
}
